% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color.R
\name{get_palette}
\alias{get_palette}
\alias{colvec}
\title{Color utilities}
\usage{
get_palette()

colvec(x)
}
\arguments{
\item{x}{Vector of values to evaluate.}
}
\value{
\code{get_palette} returns a vector of 99 hex color values. \code{colvec}
returns a hex color vector of same length as \code{x}, corresponding to
values of \code{x}.
}
\description{
From given values, generate colors for plotting.
}
\examples{
# generate data
x <- y <- seq(-3, 3, length = 30)

# color points by numeric values
plot(x, y, col=colvec(x), pch=16)

# color points by factor values
trmt <- gl(3, 10, labels = c('Q', 'R', 'S'))
plot(x, y, col=colvec(rev(trmt)), pch=16)

}
