% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color.R
\name{colvec}
\alias{colvec}
\title{Color utilities}
\usage{
colvec(x)
}
\arguments{
\item{x}{Vector of values to evaluate.}
}
\value{
Vector of color values, of same length as \code{x}.
}
\description{
From given values, generate colors for plotting.
}
\examples{
# generate data
x <- seq(-1.95, 1.95, length = 30)
y <- seq(-1.95, 1.95, length = 35)

# color points by numeric values
plot(x, y[1:30], col=colvec(x), pch=16)

# color points by factor values
trmt <- gl(3, 10, labels = c('Q', 'R', 'S'))
plot(x, y[1:30], col=colvec(rev(trmt)), pch=16)

}
